Add gitignore
The project had no gitignore and due to that had some useless system and
configuration files added to the repository. After adding the git ignore, the
project was purged from files that should have been removed from
the repository.

It's advised that all further commits should be
inspected for files that  went unnoticed in this purge.